requirements.txt
fastapi
uvicorn
requests
dnspython